trig_range_reduction: changed return value

+++ src/lib/headers/s21_fp_utils.h
@@ -20,7 +20,7 @@ bool s21_isnegative(double x);
 bool s21_ispositive(double x);
 void s21_print_bits(size_t size, void* ptr);
 void s21_print_fp_sysinfo(void);
-double s21_trig_range_reduction(double x);
+long double s21_trig_range_reduction(long double x);
 long double s21_trunc(double x);
 long double s21_factorial(int x);
